Make dedicated order-of-hooks test
